Revert "Broken Perspective"
This reverts commit 4785f2fbfa59019a1368434db6324cbdd46c7c48.

+++ Engine/src/Cobra/headers/Window.hpp
@@ -37,7 +37,7 @@ namespace Cobra
 
             void RenderView();
 
-            void Pixel(float x, float y, int r = 255, int g = 255, int b = 255);
+            void Pixel(int x, int y, int r = 255, int g = 255, int b = 255);
             bool Display();
             void CreateWindow();
             void End();
